requirements.txt
aiohttp_requests==0.1.3
discord==1.0.1
lyricsgenius==3.0.1
mysql_connector_repackaged==0.3.1
PyYAML==5.4.1
spotipy==2.18.0
youtube_dl==2021.5.16
